:lipstick: add TabNavigation.tsx
