--- src/components/layout/container/Aside.tsx
@@ -1,5 +1,0 @@
-const Aside = () => {
-  return <div></div>
-}
-
-export { Aside }
