+++ src/components/layout/container/index.ts
@@ -1,6 +1,6 @@
-export { Aside } from './Aside'
 export { AsideContextList } from './AsideContextList'
 export { Footer } from './Footer'
 export { Header } from './Header'
 export { LayoutWithHead } from './LayoutWithHead'
 export { SidebarDrawer } from './SidebarDrawer'
+export { TabNavigation } from './TabNavigation'
